reorganize AttachmentMeta
