+++ imessage-database/src/tables/messages/mod.rs
@@ -4,7 +4,6 @@
 
 pub use message::Message;
 
-pub mod attachment_metadata;
 pub(crate) mod body;
 pub mod message;
 pub mod models;
