+++ imessage-exporter/src/exporters/html.rs
@@ -34,7 +34,10 @@ use imessage_database::{
     },
     tables::{
         attachment::{Attachment, MediaType},
-        messages::{attachment_metadata::AttachmentMeta, models::BubbleComponent, Message},
+        messages::{
+            models::{AttachmentMeta, BubbleComponent},
+            Message,
+        },
         table::{Table, FITNESS_RECEIVER, ME, ORPHANED, YOU},
     },
     util::{
@@ -1581,7 +1584,7 @@ mod tests {
         app::export_type::ExportType, exporters::exporter::Writer, Config, Exporter, Options, HTML,
     };
     use imessage_database::{
-        tables::{messages::attachment_metadata::AttachmentMeta, table::ME},
+        tables::{messages::models::AttachmentMeta, table::ME},
         util::platform::Platform,
     };
